+++ src/app/(dashboard)/customers/schemas/customers-list.schema.ts
@@ -1,0 +1,9 @@
+import { z } from 'zod';
+
+import { CustomerSchema } from '@/schemas/customer.schema';
+
+export const CustomersListSchema = CustomerSchema.extend({
+  email: z.email(),
+});
+
+export type CustomersList = z.infer<typeof CustomersListSchema>;