+++ src/lib/formatters.ts
@@ -3,3 +3,15 @@ export function dateFormat(date: string | Date) {
   const formattedDate = dateFormat.toLocaleDateString('en-GB', { year: 'numeric', month: 'long', day: 'numeric' });
   return formattedDate;
 }
+
+export function dateTimeFormat(dateTime: string | Date) {
+  const dateTimeFormat = new Date(dateTime);
+  const formattedDate = dateTimeFormat.toLocaleDateString('en-GB', { year: 'numeric', month: 'long', day: 'numeric' });
+  const formattedTime = dateTimeFormat.toLocaleTimeString('en-GB', {
+    hour: 'numeric',
+    minute: 'numeric',
+    hour12: true,
+  });
+  const formattedDateTime = `${formattedDate} at ${formattedTime}`;
+  return formattedDateTime;
+}